refactor: split MomentumPointFeedback widget into animation, header, body and remove size exemption (Task P4-C)
