+++ app/lib/features/today_feed/presentation/widgets/momentum_point_feedback_widget.dart
@@ -1,4 +1,3 @@
-// @size-exempt Temporary: exceeds hard ceiling – scheduled for refactor
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import '../../../../core/theme/app_theme.dart';
@@ -6,6 +5,8 @@ import '../../../../core/services/responsive_service.dart';
 import '../../../../core/services/accessibility_service.dart';
 import '../../data/services/today_feed_momentum_award_service.dart';
 import 'dart:async';
+import 'momentum_point_feedback_body.dart';
+part 'momentum_point_feedback_animation.dart';
 
 /// Visual feedback widget for momentum point awards
 /// Displays celebratory animations and confirmation messages when users earn momentum points
@@ -134,162 +135,12 @@ class _MomentumPointFeedbackWidgetState
     super.dispose();
   }
 
-  void _setupAnimations() {
-    // Update controllers with context-dependent durations
-    _scaleController
-        .duration = ResponsiveServiceAnimations.getAnimationDuration(
-      context,
-      baseDuration: const Duration(milliseconds: 800),
-    );
+  // Delegates to extracted extension in part file
+  void _setupAnimations() => _setupAnimationsExt();
 
-    _scaleAnimation = Tween<double>(begin: 0.0, end: 1.0).animate(
-      CurvedAnimation(parent: _scaleController, curve: Curves.elasticOut),
-    );
-
-    _fadeController.duration = ResponsiveServiceAnimations.getAnimationDuration(
-      context,
-      baseDuration: const Duration(milliseconds: 600),
-    );
-
-    _fadeAnimation = Tween<double>(
-      begin: 0.0,
-      end: 1.0,
-    ).animate(CurvedAnimation(parent: _fadeController, curve: Curves.easeOut));
-
-    _slideController
-        .duration = ResponsiveServiceAnimations.getAnimationDuration(
-      context,
-      baseDuration: const Duration(milliseconds: 500),
-    );
-
-    _slideAnimation = Tween<Offset>(
-      begin: const Offset(0, 0.5),
-      end: Offset.zero,
-    ).animate(
-      CurvedAnimation(parent: _slideController, curve: Curves.easeOutCubic),
-    );
-
-    _bounceController
-        .duration = ResponsiveServiceAnimations.getAnimationDuration(
-      context,
-      baseDuration: const Duration(milliseconds: 300),
-    );
-
-    _bounceAnimation = Tween<double>(begin: 1.0, end: 1.15).animate(
-      CurvedAnimation(parent: _bounceController, curve: Curves.elasticOut),
-    );
-
-    _glowController.duration = ResponsiveServiceAnimations.getAnimationDuration(
-      context,
-      baseDuration: const Duration(milliseconds: 1200),
-    );
-
-    _glowAnimation = Tween<double>(begin: 0.0, end: 1.0).animate(
-      CurvedAnimation(parent: _glowController, curve: Curves.easeInOut),
-    );
-  }
-
-  Future<void> _startCelebrationSequence() async {
-    // Set visibility immediately for both animated and non-animated cases
-    setState(() {
-      _isVisible = true;
-    });
-
-    if (!widget.enableAnimations ||
-        AccessibilityService.shouldReduceMotion(context)) {
-      // Skip animations but still show feedback immediately
-      _fadeController.value = 1.0;
-      _scaleController.value = 1.0;
-      _slideController.value = 1.0;
-
-      if (widget.autoHide) {
-        _autoHideTimer?.cancel();
-        // Add a small delay to ensure the widget is visible for tests
-        _autoHideTimer = Timer(
-          Duration(milliseconds: widget.autoHideDuration.inMilliseconds + 50),
-          () {
-            if (mounted) {
-              _hideFeedback();
-            }
-          },
-        );
-      }
-
-      widget.onAnimationComplete?.call();
-      return;
-    }
-
-    // Haptic feedback for success
-    HapticFeedback.mediumImpact();
-
-    // Start celebration sequence with cancellable timers
-    _fadeController.forward();
-
-    _animationTimer1 = Timer(const Duration(milliseconds: 100), () {
-      if (mounted) {
-        _scaleController.forward();
-        _glowController.forward();
-      }
-    });
-
-    _animationTimer2 = Timer(const Duration(milliseconds: 200), () {
-      if (mounted) {
-        _slideController.forward();
-      }
-    });
-
-    _animationTimer3 = Timer(const Duration(milliseconds: 400), () {
-      if (mounted) {
-        _bounceController.forward().then((_) {
-          if (mounted) {
-            _bounceController.reverse();
-          }
-        });
-      }
-    });
-
-    _animationTimer4 = Timer(const Duration(milliseconds: 450), () {
-      if (mounted) {
-        HapticFeedback.lightImpact();
-      }
-    });
-
-    // Auto-hide after celebration
-    if (widget.autoHide) {
-      _autoHideTimer?.cancel();
-      _autoHideTimer = Timer(widget.autoHideDuration, () {
-        if (mounted) {
-          _hideFeedback();
-        }
-      });
-    }
-
-    widget.onAnimationComplete?.call();
-  }
-
-  Future<void> _hideFeedback() async {
-    // Cancel all timers
-    _autoHideTimer?.cancel();
-    _animationTimer1?.cancel();
-    _animationTimer2?.cancel();
-    _animationTimer3?.cancel();
-    _animationTimer4?.cancel();
-
-    if (!widget.enableAnimations ||
-        AccessibilityService.shouldReduceMotion(context)) {
-      setState(() {
-        _isVisible = false;
-      });
-      return;
-    }
-
-    await _fadeController.reverse();
-    if (mounted) {
-      setState(() {
-        _isVisible = false;
-      });
-    }
-  }
+  // Delegates to extracted extension in part file
+  Future<void> _startCelebrationSequence() async =>
+      _startCelebrationSequenceExt();
 
   @override
   Widget build(BuildContext context) {
@@ -341,205 +192,13 @@ class _MomentumPointFeedbackWidgetState
   }
 
   Widget _buildFeedbackContent(BuildContext context) {
-    if (widget.awardResult.isQueued) {
-      return RepaintBoundary(child: _buildQueuedFeedback(context));
-    }
-
-    return RepaintBoundary(child: _buildSuccessFeedback(context));
-  }
-
-  Widget _buildSuccessFeedback(BuildContext context) {
-    return Container(
-      padding: ResponsiveService.getMediumPadding(context),
-      decoration: BoxDecoration(
-        color: AppTheme.getSurfacePrimary(context),
-        borderRadius: BorderRadius.circular(
-          ResponsiveService.getBorderRadius(context),
-        ),
-        border: Border.all(
-          color: AppTheme.momentumRising.withValues(alpha: 0.3),
-          width: 2,
-        ),
-      ),
-      child: Column(
-        mainAxisSize: MainAxisSize.min,
-        children: [
-          _buildPointIndicator(context),
-          if (widget.showMessage) ...[
-            SizedBox(height: ResponsiveService.getSmallSpacing(context)),
-            _buildSuccessMessage(context),
-          ],
-        ],
-      ),
+    return MomentumPointFeedbackBody(
+      awardResult: widget.awardResult,
+      showMessage: widget.showMessage,
+      slideAnimation: _slideAnimation,
+      scaleAnimation: _scaleAnimation,
+      glowAnimation: _glowAnimation,
     );
-  }
-
-  Widget _buildQueuedFeedback(BuildContext context) {
-    return Container(
-      padding: ResponsiveService.getMediumPadding(context),
-      decoration: BoxDecoration(
-        color: AppTheme.getSurfacePrimary(context),
-        borderRadius: BorderRadius.circular(
-          ResponsiveService.getBorderRadius(context),
-        ),
-        border: Border.all(
-          color: AppTheme.momentumSteady.withValues(alpha: 0.3),
-          width: 2,
-        ),
-      ),
-      child: Column(
-        mainAxisSize: MainAxisSize.min,
-        children: [
-          Icon(
-            Icons.schedule,
-            size: ResponsiveService.getIconSize(context, baseSize: 32),
-            color: AppTheme.momentumSteady,
-          ),
-          if (widget.showMessage) ...[
-            SizedBox(height: ResponsiveService.getSmallSpacing(context)),
-            SlideTransition(
-              position: _slideAnimation,
-              child: Text(
-                'Points queued for when back online',
-                style: Theme.of(context).textTheme.bodyMedium?.copyWith(
-                  color: AppTheme.getTextPrimary(context),
-                  fontWeight: FontWeight.w600,
-                ),
-                textAlign: TextAlign.center,
-              ),
-            ),
-          ],
-        ],
-      ),
-    );
-  }
-
-  Widget _buildPointIndicator(BuildContext context) {
-    final pointSize = ResponsiveService.getIconSize(context, baseSize: 48);
-    final iconSize = ResponsiveService.getIconSize(context, baseSize: 24);
-
-    return RepaintBoundary(
-      child: AnimatedBuilder(
-        animation: _scaleAnimation,
-        builder: (context, child) {
-          return Transform.scale(
-            scale: _scaleAnimation.value,
-            child: Container(
-              width: pointSize,
-              height: pointSize,
-              decoration: const BoxDecoration(
-                shape: BoxShape.circle,
-                color: AppTheme.momentumRising,
-              ),
-              child: Stack(
-                alignment: Alignment.center,
-                children: [
-                  // Background circle with pulse effect
-                  RepaintBoundary(
-                    child: AnimatedBuilder(
-                      animation: _glowAnimation,
-                      builder: (context, child) {
-                        return Container(
-                          width: pointSize * (1 + _glowAnimation.value * 0.2),
-                          height: pointSize * (1 + _glowAnimation.value * 0.2),
-                          decoration: BoxDecoration(
-                            shape: BoxShape.circle,
-                            color: AppTheme.momentumRising.withValues(
-                              alpha: 0.3 * (1 - _glowAnimation.value),
-                            ),
-                          ),
-                        );
-                      },
-                    ),
-                  ),
-                  // Main point indicator
-                  Icon(Icons.add_circle, size: iconSize, color: Colors.white),
-                  // Point text
-                  Positioned(
-                    bottom: -2,
-                    right: -2,
-                    child: Container(
-                      padding: const EdgeInsets.symmetric(
-                        horizontal: 4,
-                        vertical: 2,
-                      ),
-                      decoration: const BoxDecoration(
-                        color: Colors.white,
-                        shape: BoxShape.circle,
-                      ),
-                      child: Text(
-                        '+${widget.awardResult.pointsAwarded}',
-                        style: Theme.of(context).textTheme.labelSmall?.copyWith(
-                          color: AppTheme.momentumRising,
-                          fontWeight: FontWeight.w800,
-                          fontSize:
-                              ResponsiveService.getFontSizeMultiplier(context) *
-                              10,
-                        ),
-                      ),
-                    ),
-                  ),
-                ],
-              ),
-            ),
-          );
-        },
-      ),
-    );
-  }
-
-  Widget _buildSuccessMessage(BuildContext context) {
-    return RepaintBoundary(
-      child: SlideTransition(
-        position: _slideAnimation,
-        child: Column(
-          children: [
-            Text(
-              'Momentum +${widget.awardResult.pointsAwarded}!',
-              style: Theme.of(context).textTheme.titleMedium?.copyWith(
-                color: AppTheme.momentumRising,
-                fontWeight: FontWeight.w700,
-              ),
-              textAlign: TextAlign.center,
-            ),
-            SizedBox(height: ResponsiveService.getTinySpacing(context)),
-            Text(
-              _getSuccessMessage(),
-              style: Theme.of(context).textTheme.bodySmall?.copyWith(
-                color: AppTheme.getTextSecondary(context),
-              ),
-              textAlign: TextAlign.center,
-              maxLines: 2,
-              overflow: TextOverflow.ellipsis,
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-
-  String _getSuccessMessage() {
-    if (widget.awardResult.isQueued) {
-      return 'Your point will be added when back online';
-    }
-
-    final messages = [
-      'Great job staying curious about your health!',
-      'Learning something new every day! 🌟',
-      'Knowledge is momentum! Keep going! 📚',
-      'Your daily dose of health wisdom! 💡',
-      'Building healthy habits, one read at a time! 🚀',
-    ];
-
-    // Use award time or current time to select a consistent message for the day
-    final seedTime = widget.awardResult.awardTime ?? DateTime.now();
-    // For deterministic testing, always return the first message if award time is null
-    // This ensures tests can predict the exact message
-    final messageIndex =
-        widget.awardResult.awardTime != null
-            ? seedTime.day % messages.length
-            : 0;
-    return messages[messageIndex];
   }
 
   String _getAccessibilityLabel() {
@@ -548,40 +207,5 @@ class _MomentumPointFeedbackWidgetState
     }
 
     return 'Momentum point awarded! You earned ${widget.awardResult.pointsAwarded} point for reading today\'s health insight';
-  }
-}
-
-/// Extension to add animation duration utility to ResponsiveService
-extension ResponsiveServiceAnimations on ResponsiveService {
-  static Duration getAnimationDuration(
-    BuildContext context, {
-    required Duration baseDuration,
-  }) {
-    if (AccessibilityService.shouldReduceMotion(context)) {
-      return Duration.zero;
-    }
-
-    // Scale animation duration based on device type for better performance
-    final deviceType = ResponsiveService.getDeviceType(context);
-    switch (deviceType) {
-      case DeviceType.mobileSmall:
-        return Duration(
-          milliseconds: (baseDuration.inMilliseconds * 0.85).round(),
-        );
-      case DeviceType.mobile:
-        return baseDuration;
-      case DeviceType.mobileLarge:
-        return Duration(
-          milliseconds: (baseDuration.inMilliseconds * 1.1).round(),
-        );
-      case DeviceType.tablet:
-        return Duration(
-          milliseconds: (baseDuration.inMilliseconds * 1.15).round(),
-        );
-      case DeviceType.desktop:
-        return Duration(
-          milliseconds: (baseDuration.inMilliseconds * 1.2).round(),
-        );
-    }
   }
 }
